Major changes to Makefiles. See entire commit msg
Installed GNU utilities for Windows, including findutils
and coreutils. In particular, access to linux `find`
allows make to build the list of modules automatically.

Also, Makefiles were updated to comply with conventions
about prerequisites (allowing make to simply exit if
targets are up-to-date -- safer in the case of .Rmd
compilation with intermediate files, if .md is modified
instead, makefile will just quit) and PHONY targets.